requirements.txt
# requirements.txt

# Spark
pyspark==3.5.3

# Presidio for PII detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Spark NLP for Named Entity Recognition (NER)
spark-nlp==5.5.1

# PyFFX for FF3 encryption (if required)


# Pandas for Pandas UDF
pandas==2.2.3
pyarrow==18.0.0

PyYAML==6.0.2